+++ skbio/alignment/_alignment.py
@@ -11,15 +11,18 @@ from future.builtins import zip, range
 from future.utils import viewkeys, viewitems
 
 from collections import Counter, defaultdict, OrderedDict
+from importlib import import_module
 from warnings import warn
 
 import numpy as np
 from scipy.stats import entropy
 
-import skbio.io
 from skbio.stats.distance import DistanceMatrix
 from skbio.io.util import open_file
 from ._exception import SequenceCollectionError, StockholmParseError
+
+# This will be the responsibility of the ABC in the future.
+import_module('skbio.io')
 
 
 class SequenceCollection(object):
@@ -958,6 +961,8 @@ class Alignment(SequenceCollection):
     <Alignment: n=2; mean +/- std length=7.00 +/- 0.00>
 
     """
+    # TODO change once we support more formats (#629)
+    default_write_format = 'phylip'
 
     def __init__(self, seqs, validate=False, score=None,
                  start_end_positions=None):
@@ -1621,10 +1626,6 @@ class Alignment(SequenceCollection):
             return 0
         else:
             return len(self._data[0])
-
-    def write(self, fp, format):
-        # TODO remove once #672 is merged
-        skbio.io.write(self, into=fp, format=format)
 
     def to_phylip(self, map_labels=False, label_prefix=""):
         """Return phylip-formatted string representing the `SequenceCollection`